Update App.java
This is going to have to be merged with Anthony's original work. The JAXB dependencies just won't work for me for some reason. As I was combing through the files, I noticed there's typos in ProjectItems.java and Item.java but only with the methods or variables named quantity, although their all misspelled the same way continuously so there's no real issue. I removed the print xml to the console and I added a button and a method that will print out the sub total of all the items in the list. There's a bug currently where it isn't catching anything that's inside of the arraylist, so this might need to be scrapped and we might have to pretend this never happened. Sorry Eli, this was a lot to read.